# Faster `intersection_array` computation for checking distance-regularity