<!--
Please use pre-commit to lint your code.
For more details check out step 1 and 4 of
https://networkx.org/documentation/latest/developer/contribute.html
-->

I made a number of improvements to the `intersection_array` function:

1. calling `nx.is_regular` instead of manually checking whether the input graph is regular;
1. the input graph is very unlikely to be distance-regular: the number b(n) of distance-regular graphs (https://oeis.org/A241814) is a tiny fraction of the number of a(n) of connected simple graphs (https://oeis.org/A001349). In light of this, computing shortest path lengths as we go instead of precomputing them all seems more reasonable.
1. checking connectedness first allows us to avoid catching exceptions in the main loop.
1. we don't need to call `nx.diameter` at the end of the function, since we can update its value as we go with simple calls to `max`.
1. moreover, keeping track of the diameter as we go also allows us to stop early: distance-regular graphs on n vertices are known to have a diameter of at most (8 * log_2 n) / 3 (see https://doi.org/10.1016/j.ejc.2004.07.004, Theorem 1.5 page 81), so if at any point the current diameter's value exceeds that then we can give up early.
1. no need to build lists for computing `b` and `c`, since we are only interested in cardinality.
1. I thought bidirectional searches for shortest (u, v)-paths would (asymptotically) amount to more work than launching single-source shortest paths algorithms, since the former uses two BFS anyway. After comparing both approaches in practice, I turned out to be right: the bidirectional searches often take much longer than the original version of `intersection_array`.

# Results on distance-regular graphs

I compared the running times of the current version of `intersection_array` with mine on the ten largest distance-regular graphs I could get (data comes from https://www.distanceregular.org/). `is_distance_regular` is NetworkX's current implementation, `my_is_distance_regular` uses the proposed improvement, and `my_intersection_array_2` computes (u, v)-paths rather than the single source paths to all vertices.

```
Comparing running times for distance regularity checking

The following instances are known to be distance-regular
now dealing with instance  ./tests/test_data/distance-regular/coset-externarygolay.s6             729 vertices, 8748 edges
    now running is_distance_regular     done in 2.1808555030147545 seconds, answer is True
    now running my_is_distance_regular  done in 1.7023403490020428 seconds, answer is True
    now running my_intersection_array_2 done in 4.411169429979054 seconds, answer is ([24, 22, 20], [1, 2, 12])

now dealing with instance  ./tests/test_data/distance-regular/shortened-extended-ternary-golay.s6 729 vertices, 8019 edges
    now running is_distance_regular     done in 3.3089937710028607 seconds, answer is True
    now running my_is_distance_regular  done in 2.775795002002269 seconds, answer is True
    now running my_intersection_array_2 done in 4.316444240976125 seconds, answer is ([22, 20, 18, 2, 1], [1, 2, 9, 20, 22])

now dealing with instance  ./tests/test_data/distance-regular/games.s6                            729 vertices, 40824 edges
    now running is_distance_regular     done in 7.66312413700507 seconds, answer is True
    now running my_is_distance_regular  done in 6.080848899000557 seconds, answer is True
    now running my_intersection_array_2 done in 8.80728219798766 seconds, answer is ([112, 110], [1, 20])

now dealing with instance  ./tests/test_data/distance-regular/witt.s6                             759 vertices, 11385 edges
    now running is_distance_regular     done in 2.950715120008681 seconds, answer is True
    now running my_is_distance_regular  done in 2.4445975279959384 seconds, answer is True
    now running my_intersection_array_2 done in 5.15949073000229 seconds, answer is ([30, 28, 24], [1, 3, 15])

now dealing with instance  ./tests/test_data/distance-regular/iif.s6                              990 vertices, 3465 edges
    now running is_distance_regular     done in 2.6043268000066746 seconds, answer is True
    now running my_is_distance_regular  done in 2.290155616006814 seconds, answer is True
    now running my_intersection_array_2 done in 11.459608876990387 seconds, answer is ([7, 6, 4, 4, 4, 1, 1, 1], [1, 1, 1, 2, 4, 4, 6, 7])

now dealing with instance  ./tests/test_data/distance-regular/sksgraph.s6                         1024 vertices, 16896 edges
    now running is_distance_regular     done in 5.97368586299126 seconds, answer is True
    now running my_is_distance_regular  done in 5.078227958001662 seconds, answer is True
    now running my_intersection_array_2 done in 9.856664725986775 seconds, answer is ([33, 30, 15], [1, 2, 15])

now dealing with instance  ./tests/test_data/distance-regular/cube10.s6                           1024 vertices, 5120 edges
    now running is_distance_regular     done in 3.033381476998329 seconds, answer is True
    now running my_is_distance_regular  done in 2.5856517939828336 seconds, answer is True
    now running my_intersection_array_2 done in 30.335086243983824 seconds, answer is ([10, 9, 8, 7, 6, 5, 4, 3, 2, 1], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10])

now dealing with instance  ./tests/test_data/distance-regular/hamming4.6.s6                       1296 vertices, 12960 edges
    now running is_distance_regular     done in 6.618309289013268 seconds, answer is True
    now running my_is_distance_regular  done in 5.388119548995746 seconds, answer is True
    now running my_intersection_array_2 done in 28.552194002026226 seconds, answer is ([20, 15, 10, 5], [1, 2, 3, 4])

now dealing with instance  ./tests/test_data/distance-regular/grassmann263.s6                     1395 vertices, 68355 edges
    now running is_distance_regular     done in 32.80391989799682 seconds, answer is True
    now running my_is_distance_regular  done in 24.02292016299907 seconds, answer is True
    now running my_intersection_array_2 done in 39.16650592099177 seconds, answer is ([98, 72, 32], [1, 9, 49])

now dealing with instance  ./tests/test_data/distance-regular/odd7.s6                             1716 vertices, 6006 edges
    now running is_distance_regular     done in 7.135661044012522 seconds, answer is True
    now running my_is_distance_regular  done in 6.108615503995679 seconds, answer is True
    now running my_intersection_array_2 done in 54.69806831300957 seconds, answer is ([7, 6, 6, 5, 5, 4], [1, 1, 2, 2, 3, 3])

Comparing totals:
    is_distance_regular    : 74.27297290205024
    my_is_distance_regular : 58.477272361982614
    my_intersection_array_2: 196.76251468193368
```
    
# Results on random regular graphs

Same as above, but with random regular graphs.
    
```
The following 10 instances random 20-regular graphs
now dealing with instance 0 -- 2048 vertices, 20480 edges
    now running is_distance_regular     done in 2.2792291679943446 seconds, answer is False
    now running my_is_distance_regular  done in 0.0033934070088434964 seconds, answer is False
    now running my_intersection_array_2 done in 0.002207571000326425 seconds, answer is False

now dealing with instance 1 -- 2048 vertices, 20480 edges
    now running is_distance_regular     done in 2.315355560014723 seconds, answer is False
    now running my_is_distance_regular  done in 0.003575303009711206 seconds, answer is False
    now running my_intersection_array_2 done in 0.0019738929986488074 seconds, answer is False

now dealing with instance 2 -- 2048 vertices, 20480 edges
    now running is_distance_regular     done in 2.2876819689990953 seconds, answer is False
    now running my_is_distance_regular  done in 0.003264581988332793 seconds, answer is False
    now running my_intersection_array_2 done in 0.001720152999041602 seconds, answer is False

now dealing with instance 3 -- 2048 vertices, 20480 edges
    now running is_distance_regular     done in 2.2935480609885417 seconds, answer is False
    now running my_is_distance_regular  done in 0.003167785005643964 seconds, answer is False
    now running my_intersection_array_2 done in 0.002119503973517567 seconds, answer is False

now dealing with instance 4 -- 2048 vertices, 20480 edges
    now running is_distance_regular     done in 2.2664987099997234 seconds, answer is False
    now running my_is_distance_regular  done in 0.0032383989891968668 seconds, answer is False
    now running my_intersection_array_2 done in 0.00220975698903203 seconds, answer is False

now dealing with instance 5 -- 2048 vertices, 20480 edges
    now running is_distance_regular     done in 2.3034619920072146 seconds, answer is False
    now running my_is_distance_regular  done in 0.0033776959753595293 seconds, answer is False
    now running my_intersection_array_2 done in 0.0020072120241820812 seconds, answer is False

now dealing with instance 6 -- 2048 vertices, 20480 edges
    now running is_distance_regular     done in 2.36016765399836 seconds, answer is False
    now running my_is_distance_regular  done in 0.0032529000018257648 seconds, answer is False
    now running my_intersection_array_2 done in 0.002387827989878133 seconds, answer is False

now dealing with instance 7 -- 2048 vertices, 20480 edges
    now running is_distance_regular     done in 2.5374458819860592 seconds, answer is False
    now running my_is_distance_regular  done in 0.00350969098508358 seconds, answer is False
    now running my_intersection_array_2 done in 0.002192362997448072 seconds, answer is False

now dealing with instance 8 -- 2048 vertices, 20480 edges
    now running is_distance_regular     done in 2.6173524199984968 seconds, answer is False
    now running my_is_distance_regular  done in 0.004194754990749061 seconds, answer is False
    now running my_intersection_array_2 done in 0.0025923799839802086 seconds, answer is False

now dealing with instance 9 -- 2048 vertices, 20480 edges
    now running is_distance_regular     done in 2.5984127949923277 seconds, answer is False
    now running my_is_distance_regular  done in 0.0035534100024960935 seconds, answer is False
    now running my_intersection_array_2 done in 0.002110676985466853 seconds, answer is False

Comparing totals:
    is_distance_regular    : 23.859154210978886
    my_is_distance_regular : 0.034527927957242355
    my_intersection_array_2: 0.02152133794152178

```