@alabarre 2025-07-31
Hi Dan,

I approved your changes, I think we can merge the current version and process further optimizations in a later PR.

Thank you all!